+++ CHANGELOG.md
@@ -22,6 +22,16 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+- Support direct local connection to JDownloader using their deprecated API
+
+### Changed
+
+- Log HTTP requests made to JDownloader
+
 ## [10.5.0] - 2026-08-15
 
 ### Added